A version that can work in Windows with bugs.
This version is still with bugs.

Network-related functions are not enabled.

Some header files are divided into a .h file with a .cc file.

No test is done in Linux for this version.

A README.md file is added for introducing the file structure and requirements in Windows.

+++ src/core/binary_counter.h
@@ -13,77 +13,77 @@
 
 namespace qx {
 
-   class binary_counter
-   {
-      private:
+    class binary_counter
+    {
+    private:
 
-	 uint64_t __val;
-	 uint64_t __max;
-	 uint64_t __num;
+        uint64_t __val;
+        uint64_t __max;
+        uint64_t __num;
 
-      public:
+    public:
 
-	 binary_counter(uint64_t bit_count)
-	 {
-	    assert(bit_count <= 64);
-	    __num = bit_count;
-	    __max = 1 << bit_count;
-	    __val = 0;
-	 }
+        binary_counter(uint64_t bit_count)
+        {
+            assert(bit_count <= 64);
+            __num = bit_count;
+            __max = static_cast<uint64_t>(1 << bit_count);
+            __val = 0;
+        }
 
-	 void set(uint32_t b)
-	 {
-	    __val |= (1 << b);
-	 }
+        void set(uint32_t b)
+        {
+            __val |= static_cast<uint64_t>(1 << b);
+        }
 
-	 void unset(uint32_t b)
-	 {
-	    __val &= ~(1 << b);
-	 }
+        void unset(uint32_t b)
+        {
+            __val &= ~(1 << b);
+        }
 
-	 uint64_t test(uint32_t b)
-	 {
-	    return (__val & (1 << b));
-	 }
+        uint64_t test(uint32_t b)
+        {
+            return (__val & static_cast<uint64_t>(1 << b));
+        }
 
-	 void reset()
-	 {
-	    __val = 0;
-	 }
+        void reset()
+        {
+            __val = 0;
+        }
 
-	 uint64_t value()
-	 {
-	    return __val;
-	 }
+        uint64_t value()
+        {
+            return __val;
+        }
 
-	 void dump()
-	 {
-	    print("[ ");
-	    for (uint32_t i=__num; i>0;i--)
-	       print((test(i-1) ? '1' : '0'));
-	    println(" : " << __val << "]");
-	 }
+        void dump()
+        {
+            print("[ ");
+            for (uint64_t i = __num; i > 0; i--)
+                print((test(static_cast<uint32_t>(i - 1)) ? '1' : '0'));
+            println(" : " << __val << "]");
+        }
 
-	 binary_counter& operator=(uint64_t v)
-	 {
-	    __val = v;
-	    return *this;
-	 }
+        binary_counter& operator=(uint64_t v)
+        {
+            __val = v;
+            return *this;
+        }
 
-	 binary_counter& operator++()
-	 {
-	    __val++;
-	    return *this;
-	 }
+        binary_counter& operator++()
+        {
+            __val++;
+            return *this;
+        }
 
-	 binary_counter operator++(int) 
-	 {
-	    binary_counter tmp(*this); // copy
-	    operator++(); // pre-increment
-	    return tmp;   // return old value
-	 }
+        binary_counter operator++(int)
+        {
+            binary_counter tmp(*this); // copy
+            operator++(); // pre-increment
+            return tmp;   // return old value
+        }
 
-   };
+    };
 } // qx
